% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{A}
\alias{A}
\alias{aa}
\alias{AZ}
\alias{az}
\title{The elliptical arc curve commands}
\usage{
A(
  rx,
  ry = rx,
  x_axis_rotation = 0,
  large_arc_flag = FALSE,
  sweep_flag = FALSE,
  x,
  y = NULL,
  sep = getOption("dee.sep", ",")
)

aa(
  rx,
  ry = rx,
  x_axis_rotation = 0,
  large_arc_flag = FALSE,
  sweep_flag = FALSE,
  x,
  y = NULL,
  sep = getOption("dee.sep", ",")
)

AZ(...)

az(...)
}
\arguments{
\item{rx, ry}{Radius of ellipse.}

\item{x_axis_rotation}{Angle (in degrees) from x-axis of ellipse.}

\item{large_arc_flag}{If \code{TRUE} then one of two larger arc sweeps chosen else
one of the two smaller arc sweeps.}

\item{sweep_flag}{If \code{TRUE} then arc will be drawn in "positive-angle" direction.
else drawn in "negative-angle" direction.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{...}{Passed to related function.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{A()} and \code{aa()} draw elliptical arc curve commands.
}
