% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dee.r
\name{as_omsvg}
\alias{as_omsvg}
\title{Convert to an \code{omsvg} "svg" class object}
\usage{
as_omsvg(
  x,
  ...,
  height = getOption("dee.height"),
  width = getOption("dee.width"),
  background_color = getOption("dee.background_color", "none"),
  stroke = getOption("dee.stroke"),
  stroke_width = getOption("dee.stroke_width"),
  fill = getOption("dee.fill"),
  attrs = getOption("dee.attrs")
)
}
\arguments{
\item{x}{An object of class "x".}

\item{...}{Passed to \code{\link[omsvg:svg_path]{omsvg::svg_path()}}.}

\item{height, width}{svg width and height.}

\item{background_color}{Background color.}

\item{stroke, stroke_width, fill, attrs}{Passed to \code{\link[omsvg:svg_path]{omsvg::svg_path()}}.}
}
\description{
\code{as_omsvg()} converts a \code{dee} object to an \code{omsvg} "svg" class object
using \code{\link[omsvg:SVG]{omsvg::SVG()}} and \code{\link[omsvg:svg_path]{omsvg::svg_path()}}.
}
\examples{
s <- affiner::star_inner_radius(n = 5, d = 2)
d <- d_star(x = 50, y = 50, r = 30, s = s, n = 5, digits = 0)
if (requireNamespace("omsvg", quietly = TRUE)) {
  svg <- as_omsvg(d, height = 100, width = 100,
                fill = "fill", stroke = "none")
  paste(svg, collapse = "\n") |> cat("\n")
}
}
