% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{C}
\alias{C}
\alias{cc}
\alias{CZ}
\alias{cz}
\alias{S}
\alias{ss}
\alias{SZ}
\alias{sz}
\title{The cubic Bezier curve commands}
\usage{
C(
  x1,
  y1 = NULL,
  x2,
  y2 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

cc(
  x1,
  y1 = NULL,
  x2,
  y2 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

CZ(...)

cz(...)

S(
  x2,
  y2 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

ss(
  x2,
  y2 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

SZ(...)

sz(...)
}
\arguments{
\item{x1}{If \code{y1} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y1}{Either \code{NULL} or a numeric vector.}

\item{x2}{If \code{y2} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y2}{Either \code{NULL} or a numeric vector.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by \code{height - y}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{C()} and \code{cc()} draw cubic Bezier curves
\code{S()} and \code{ss()} draw cubic Bezier curves assuming the first control point is the reflection of the previous Bezier curve command.
}
\examples{
M(1, 1) + C(2, 2, 3, 3, 4, 4) + Z()
M(1, 1) + cc(1, 1, 2, 2, 3, 3) + zz()
}
