% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{CIRCLE}
\alias{CIRCLE}
\title{Circle path convenience wrapper}
\usage{
CIRCLE(x, y = NULL, r, ...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{r}{The radius of the circle.}

\item{...}{Passed to \code{\link[=M]{M()}} and \code{\link[=AZ]{AZ()}}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{MAZ_CIRCLE()} is a wrapper around \code{M()} and \code{AZ()} to create
circle shaped paths.
}
\examples{
CIRCLE(x = 5, y = 5, r = 2)
}
