% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{d_aabb}
\alias{d_aabb}
\title{Axis-aligned bounding box convenience wrapper}
\usage{
d_aabb(x, y = NULL, ...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to \code{\link[=d_polygon]{d_polygon()}}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{d_aabb()} is a wrapper around \code{d_rect()} to create
an axis-aligned bounding box path.
}
\examples{
d_aabb(x = 2:6, y = 4:8)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d_aabb(x = 2:6, y = 4:8),
       height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[=d_rect]{d_rect()}}, \code{\link[=d_polygon]{d_polygon()}}
}
