% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{d_ellipse}
\alias{d_ellipse}
\alias{d_circle}
\title{Ellipse path convenience wrapper}
\usage{
d_ellipse(
  x,
  y = NULL,
  rx,
  ry = rx,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)

d_circle(x, y = NULL, r, ...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{rx, ry}{The radii of the ellipse.}

\item{a}{The angle of rotation (will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}).}

\item{...}{Passed to \code{\link[=M]{M()}} and \code{\link[=AZ]{AZ()}}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}

\item{r}{The radius of the circle.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{d_ellipse()} is a wrapper around \code{M()} and \code{AZ()} to create
ellipse shaped paths.
It's vectorized in its \code{x}, \code{y}, \code{rx}, and \code{ry} arguments.
\code{d_circle()} is a special case to create circle shaped paths.
It's vectorized in its \code{x}, \code{y}, and \code{r} arguments.
}
\examples{
d_circle(x = 5, y = 5, r = 2)
d_ellipse(x = 5, y = 5, rx = 2, ry = 3)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d_ellipse(x = 5, y = 5, rx = 2, ry = 3),
       height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
