% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotoxal_2ngon.R
\name{d_isotoxal_2ngon}
\alias{d_isotoxal_2ngon}
\alias{d_star}
\title{Isotoxal \verb{2n}-gon path convenience wrapper}
\usage{
d_isotoxal_2ngon(
  x,
  y,
  r,
  n,
  a = 90,
  ...,
  s = affiner::isotoxal_2ngon_inner_radius(n, alpha = alpha, beta_ext = beta_ext, d = d),
  alpha = NULL,
  beta_ext = NULL,
  d = NULL,
  offset = 0,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)

d_star(
  x,
  y,
  r,
  n,
  a = 90,
  ...,
  s = affiner::isotoxal_2ngon_inner_radius(n, alpha = alpha, beta_ext = beta_ext, d = d),
  alpha = NULL,
  beta_ext = NULL,
  d = NULL,
  offset = 0,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{r}{The outer radius of the isotoxal \verb{2n}-gon polygon.}

\item{n}{The number of outer vertices.}

\item{a}{The angle of the first outer vertex (will be coerced by
\code{\link[affiner:degrees]{affiner::degrees()}}).}

\item{...}{Passed to \code{\link[=d_polygon]{d_polygon()}}.}

\item{s}{The inner radius of the isotoxal \verb{2n}-gon polygon as a
\strong{fraction} of the outer radius i.e. the inner radius = \code{s * r}.
Defaults to \code{\link[affiner:isotoxal_2ngon_inner_radius]{affiner::isotoxal_2ngon_inner_radius()}} computed from
exactly one of \code{alpha}, \code{beta_ext}, or \code{d}.}

\item{alpha}{Interior angle of an outer vertex passed to
\code{\link[affiner:isotoxal_2ngon_inner_radius]{affiner::isotoxal_2ngon_inner_radius()}} to compute \code{s}.
Will be coerced by \code{\link[affiner:degrees]{affiner::degrees()}}.}

\item{beta_ext}{Exterior angle of an inner vertex passed to
\code{\link[affiner:isotoxal_2ngon_inner_radius]{affiner::isotoxal_2ngon_inner_radius()}} to compute \code{s}.
Will be coerced by \code{\link[affiner:degrees]{affiner::degrees()}}.}

\item{d}{Density (winding number) of the star polygon \verb{|n/d|} passed to
\code{\link[affiner:isotoxal_2ngon_inner_radius]{affiner::isotoxal_2ngon_inner_radius()}} to compute \code{s}.}

\item{offset}{If a positive number the distance for \emph{outward} polygon offsetting.  If a negative number the distance for \emph{inward} polygon offsetting.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{d_isotoxal_2ngon()} is a wrapper around \code{d_polygon()}
to create isotoxal \verb{2n}-gon polygon shaped paths.
Its vectorized in its \code{x}, \code{y}, \code{r}, \code{n}, \code{a}, \code{s}, and \code{offset} arguments.
\code{d_star()} is provided as an alias.
}
\examples{
# A |5/2| star e.g. the *verda stelo* (using `d` shortcut)
d <- d_isotoxal_2ngon(x = 5, y = 5, r = 3, n = 5, d = 2)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "green", stroke = "none")
}
# "lozenge" shape most often has acute angles of 45 degrees (using `alpha`)
d <- d_isotoxal_2ngon(x = 5, y = 5, r = 4, n = 2, alpha = 45)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "cyan", stroke = "black", stroke_width = 4)
}
# `d_star()` is an alias for `d_isotoxal_2ngon()`
# Inner exterior angle of |8/3| star is 90 degrees (using `beta_ext`)
d <- d_star(x = 5, y = 5, r = 4, n = 8, a = 22.5, beta_ext = 90)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "yellow", stroke = "black", stroke_width = 4)
}

# Degenerate case of inner vertex with exterior angle of 180 degrees
# creates a regular `n`-gon.
d <- d_star(x = 5, y = 5, r = 4, n = 8, a = 22.5, beta_ext = 180)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[=d_polygon]{d_polygon()}} and \code{\link[=d_regular_ngon]{d_regular_ngon()}}.  See
\url{https://en.wikipedia.org/wiki/Isotoxal_figure#Isotoxal_polygons} and
\url{https://en.wikipedia.org/wiki/Star_polygon#Isotoxal_star_simple_polygons}
for more information on isotoxal polygons.
}
