% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{d_rect}
\alias{d_rect}
\title{Rectangle path convenience wrapper}
\usage{
d_rect(
  x,
  y = NULL,
  w,
  h,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{w, h}{The width and height of the rectangle.}

\item{a}{The angle of rotation (will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}).}

\item{...}{Passed to \code{\link[=d_polygon]{d_polygon()}}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{d_rect()} is a wrapper around \code{d_polygon()} to create
rectangle shaped paths.
It's vectorized in its \code{x}, \code{y}, \code{w}, and \code{h} arguments.
}
\examples{
d_rect(x = 5, y = 5, w = 4, h = 6)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d_rect(x = 5, y = 5, w = 4, h = 6),
       height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[=d_aabb]{d_aabb()}}, \code{\link[=d_polygon]{d_polygon()}}
}
