% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dee.r
\name{dee}
\alias{dee}
\title{Build an object of class "dee"}
\usage{
dee(x)
}
\arguments{
\item{x}{A character vector of length 1.}
}
\value{
An object of class "dee".
}
\description{
\code{dee()} casts a string to an object of class "dee".
}
\examples{
dee("M 10,30") + dee("V 30")
}
