% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dee_options.r
\name{dee_options}
\alias{dee_options}
\title{Get dee options}
\usage{
dee_options(..., default = FALSE)
}
\arguments{
\item{...}{\code{dee} package options using \code{name = value}.
The return list will use any of these instead of the current/default values.}

\item{default}{If \code{TRUE} return the default values instead of current values.}
}
\value{
A list of option values.
Note this function \strong{does not} set option values itself but
this list can be passed to \code{\link[=options]{options()}}, \code{\link[withr:with_options]{withr::local_options()}}, or \code{\link[withr:with_options]{withr::with_options()}}.
}
\description{
\code{dee_options()} returns the \code{dee} package's global options.
}
\examples{
  dee_options()

  dee_options(default = TRUE)

  dee_options(dee.height = 100, dee.width = 100, dee.origin_at_bottom = TRUE)
}
\seealso{
\link{dee-package} for a high-level description of relevant global options.
}
