% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflacionar}
\alias{deflacionar}
\alias{deflacionar.default}
\title{Deflacionar serie de valores em reais}
\usage{
deflacionar(x, ...)

\method{deflacionar}{default}(x, periodo, base, indice = "ipca", ...)
}
\arguments{
\item{x}{Valores}

\item{...}{Outros argumentos passados para os métodos}

\item{periodo}{Período de referência dos valores de \code{x}}

\item{base}{Período base dos valores finais}

\item{indice}{Índice usado para deflacionar}
}
\value{
Um vetor com valores deflacionados
}
\description{
Deflacionar serie de valores em reais
}
\examples{
serie_dbl <- c(100, 105)
datas <- seq(as.Date("2019-01-01"), by = "1 month", along.with = serie_dbl)
deflacionar(serie_dbl, datas, "2019-01-01")
}
