% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{desazonalizar}
\alias{desazonalizar}
\title{Dessazonalizar uma serie}
\usage{
desazonalizar(x, start, freq)
}
\arguments{
\item{x}{Série para dessazonalizar}

\item{start}{Início da série}

\item{freq}{Frequência da série}
}
\value{
\code{x} sem sazonalidade
}
\description{
Dessazonalizar uma serie
}
