% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{ELLIPSE}
\alias{ELLIPSE}
\alias{CIRCLE}
\title{Ellipse path convenience wrapper}
\usage{
ELLIPSE(x, y = NULL, rx, ry = rx, ...)

CIRCLE(x, y = NULL, r, ...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{rx, ry}{The radii of the ellipse.}

\item{...}{Passed to \code{\link[=M]{M()}} and \code{\link[=AZ]{AZ()}}.}

\item{r}{The radius of the circle.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{ELLIPSE()} is a wrapper around \code{M()} and \code{AZ()} to create
ellipse shaped paths.
It's vectorized in its \code{x}, \code{y}, \code{rx}, and \code{ry} arguments.
\code{CIRCLE()} is a special case to create circle shaped paths.
It's vectorized in its \code{x}, \code{y}, and \code{r} arguments.
}
\examples{
CIRCLE(x = 5, y = 5, r = 2)
ELLIPSE(x = 5, y = 5, rx = 2, ry = 3)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(ELLIPSE(x = 5, y = 5, rx = 2, ry = 3),
       height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
