% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slash.R
\name{height_slash_left}
\alias{height_slash_left}
\alias{height_slash_right}
\alias{width_slash_left}
\alias{width_slash_right}
\title{Height/width of slash ends}
\usage{
height_slash_left(
  dx,
  dy,
  w,
  ...,
  nib = c("horizontal", "diagonal", "square", "vertical"),
  left = nib,
  right = nib
)

height_slash_right(
  dx,
  dy,
  w,
  ...,
  nib = c("horizontal", "diagonal", "square", "vertical"),
  left = nib,
  right = nib
)

width_slash_left(
  dx,
  dy,
  w,
  ...,
  nib = c("horizontal", "diagonal", "square", "vertical"),
  left = nib,
  right = nib
)

width_slash_right(
  dx,
  dy,
  w,
  ...,
  nib = c("horizontal", "diagonal", "square", "vertical"),
  left = nib,
  right = nib
)
}
\arguments{
\item{dx}{The difference \code{x_right - x_left} in \code{\link[=d_fslash]{d_fslash()}} and \code{\link[=d_bslash]{d_bslash()}}.}

\item{dy}{The difference \code{y_bottom - y_top} in \code{\link[=d_fslash]{d_fslash()}} and \code{\link[=d_bslash]{d_bslash()}}.}

\item{w}{The (stroke) width of the slash "line".}

\item{...}{Must be empty.}

\item{nib, left, right}{The shape of the "nib" tracing the line.
This only affects the left and right ends.}
}
\value{
A numeric value.
}
\description{
\code{height_slash_left()}, \code{height_slash_right()},
\code{width_slash_left()}, and \code{width_slash_right()}
return the height or width of the left or right end of the equivalent
\code{\link[=d_fslash]{d_fslash()}} or \code{\link[=d_bslash]{d_bslash()}} shapes.
}
\examples{
width_slash_left(6, 6, 1, nib = "horizontal")
width_slash_left(6, 6, 1, left = "horizontal", right = "vertical")
height_slash_left(6, 6, 1, nib = "diagonal")
height_slash_right(6, 6, 1, nib = "diagonal")
width_slash_right(6, 6, 1, nib = "vertical")
}
\seealso{
\code{\link[=d_fslash]{d_fslash()}} and \code{\link[=d_bslash]{d_bslash()}}
}
