% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indice}
\alias{indice}
\alias{indice_movel}
\alias{mudar_base}
\title{Criar indice de base fixa}
\usage{
indice(x)

indice_movel(x)

mudar_base(x, base)
}
\arguments{
\item{x}{Valores}
}
\value{
Um vetor
}
\description{
Criar indice de base fixa
}
\examples{
indice(c(25, 30, 35))
}
