% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotoxal_2ngon.r
\name{ISOTOXAL_2NGON}
\alias{ISOTOXAL_2NGON}
\title{Isotoxal \verb{2n}-gon path convenience wrapper}
\usage{
ISOTOXAL_2NGON(
  x,
  y,
  r,
  s,
  n,
  a = 90,
  ...,
  offset = 0,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{r}{The outer radius of the isotoxal \verb{2n}-gon polygon.}

\item{s}{The inner radius of the isotoxal \verb{2n}-gon polygon as a \strong{fraction} of the outer radius i.e. the inner radius = \code{s * r}.  \code{\link[affiner:isotoxal_2ngon_inner_radius]{affiner::isotoxal_2ngon_inner_radius()}} may be of help computing this.}

\item{n}{The number of outer vertices.}

\item{a}{The angle of the first outer vertex (will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}).}

\item{...}{Passed to \code{\link[=POLYGON]{POLYGON()}}.}

\item{offset}{If a positive number the distance for \emph{outward} polygon offsetting.  If a negative number the distance for \emph{inward} polygon offsetting.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{ISOTOXAL_2NGON()} is a wrapper around \code{POLYGON()}
to create isotoxal \verb{2n}-gon polygon shaped paths.
Its vectorized in its \code{x}, \code{y}, \code{r}, \code{s}, \code{n}, \code{a}, and \code{offset} arguments.
}
\examples{
# A |5/2| star e.g. the *verda stelo*
s <- affiner::isotoxal_2ngon_inner_radius(n = 5, d = 2)
d <- ISOTOXAL_2NGON(x = 5, y = 5, r = 3, s = s, n = 5)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "green", stroke = "none")
}
# "lozenge" shape most often has acute angles of 45 degrees
s <- affiner::isotoxal_2ngon_inner_radius(n = 2, alpha = 45)
d <- ISOTOXAL_2NGON(x = 5, y = 5, r = 4, s = s, n = 2)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "cyan", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[=POLYGON]{POLYGON()}} and \code{\link[=REGULAR_NGON]{REGULAR_NGON()}}.  See \url{https://en.wikipedia.org/wiki/Isotoxal_figure#Isotoxal_polygons} and \url{https://en.wikipedia.org/wiki/Star_polygon#Isotoxal_star_simple_polygons} for more information on isotoxal polygons.
}
