% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{participacao}
\alias{participacao}
\title{Calcular partipacao com base em quantidades}
\usage{
participacao(x)
}
\arguments{
\item{x}{vetor de quantidades}
}
\value{
um vetor de participacoes
}
\description{
Calcular partipacao com base em quantidades
}
\examples{
participacao(c(25, 25))
}
