% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dee.r
\name{plot.dee}
\alias{plot.dee}
\title{Plot an object of class "dee"}
\usage{
\method{plot}{dee}(
  x,
  ...,
  height = getOption("dee.height"),
  width = getOption("dee.width"),
  background_color = getOption("dee.background_color", "none"),
  stroke = getOption("dee.stroke"),
  stroke_width = getOption("dee.stroke_width"),
  fill = getOption("dee.fill"),
  attrs = getOption("dee.attrs")
)
}
\arguments{
\item{x}{An object of class "x".}

\item{...}{Passed to \code{\link[omsvg:svg_path]{omsvg::svg_path()}}.}

\item{height, width}{svg width and height.}

\item{background_color}{Background color.}

\item{stroke, stroke_width, fill, attrs}{Passed to \code{\link[omsvg:svg_path]{omsvg::svg_path()}}.}
}
\value{
\code{invisible(NULL)}
}
\description{
\code{plot()} an object of class "dee" using \code{\link[omsvg:SVG]{omsvg::SVG()}}, \code{\link[omsvg:svg_path]{omsvg::svg_path()}}, and \code{svgparser::read_svg()}.
}
\details{
This function requires the package \code{svgparser} which (as of January 2026) is not available on CRAN.  You can install it with \code{remotes::install_github('coolbutuseless/svgparser')} or \code{utils::install.packages('svgparser', repos = c('https://trevorld.r-universe.dev', 'https://cloud.r-project.org'))}.
}
\examples{
d <- M(10, 30) +
     A(20, 20, 0, 0, 1, 50, 30) +
     A(20, 20, 0, 0, 1, 90, 30) +
     Q(90, 60, 50, 90) +
     Q(10, 60, 10, 30) +
     Z()
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 100, width = 100, background_color = "cyan",
       fill = "red", stroke = "black", stroke_width = 4)
}
}
