% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{POLYGON}
\alias{POLYGON}
\title{Polygon path convenience wrapper}
\usage{
POLYGON(
  x,
  y = NULL,
  ...,
  offset = 0,
  linejoin = c("miter", "round"),
  miterlimit = 4
)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to \code{\link[=MZ]{MZ()}}.}

\item{offset}{If a positive number the distance for \emph{outward} polygon offsetting.  If a negative number the distance for \emph{inward} polygon offsetting.}

\item{linejoin}{If \code{offset} is nonzero the type of join operation
to use at each vertex when computing the offset region.}

\item{miterlimit}{Tolerance parameter if \code{offset} is nonzero
and \code{linejoin = "miter"}.  See \code{\link[polyclip:polyoffset]{polyclip::polyoffset()}}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{POLYGON()} is a wrapper around \code{MZ()} to create
polygon shaped paths.
If the argument \code{offset} is nonzero
will use \code{\link[polyclip:polyoffset]{polyclip::polyoffset()}} to compute an offset region.
It's vectorized in its \code{offset} argument.
}
\examples{
l <- list(x = c(2, 5, 8, 5), y = c(5, 8, 5, 2))
d <- POLYGON(l, offset = c(1, 0, -1))
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       attrs = list(fill_rule = "evenodd"),
       fill = "red", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[polyclip:polyoffset]{polyclip::polyoffset()}} for details on computing the offset region when \code{offset} is non-zero.
}
