% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_wrappers.r
\name{RECT}
\alias{RECT}
\title{Rectangle path convenience wrapper}
\usage{
RECT(x, y = NULL, w, h, ...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{w, h}{The width and height of the rectangle.}

\item{...}{Passed to \code{\link[=MZ]{MZ()}}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{RECT()} is a wrapper around \code{MZ()} to create
rectangle shaped paths.
It's vectorized in its \code{x}, \code{y}, \code{w}, and \code{h} arguments.
}
\examples{
RECT(x = 5, y = 5, w = 4, h = 6)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(RECT(x = 5, y = 5, w = 4, h = 6),
       height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
