% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotoxal_2ngon.r
\name{REGULAR_NGON}
\alias{REGULAR_NGON}
\title{Regular \code{n}-gon path convenience wrapper}
\usage{
REGULAR_NGON(
  x,
  y,
  r,
  n,
  a = 90,
  ...,
  offset = 0,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height")
)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{r}{The radius of the regular polygon (where the vertices lay on).}

\item{n}{The number of vertices.}

\item{a}{The angle of the first vertex (will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}).}

\item{...}{Passed to \code{\link[=POLYGON]{POLYGON()}}.}

\item{offset}{If a positive number the distance for \emph{outward} polygon offsetting.  If a negative number the distance for \emph{inward} polygon offsetting.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{REGULAR_NGON()} is a wrapper around \code{POLYGON()}
to create regular polygon shaped paths.
Its vectorized in its \code{x}, \code{y}, \code{r}, \code{n}, \code{a}, and \code{offset} arguments.
}
\examples{
# A pentagon
d5 <- REGULAR_NGON(x = 5, y = 5, r = 4, n = 5)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d5, height = 10, width = 10,
       fill = "magenta", stroke = "black", stroke_width = 4)
}
# A hexagon
d6 <- REGULAR_NGON(x = 5, y = 5, r = 4, n = 6, offset = c(0, -1))
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d6, height = 10, width = 10,
       attrs = list(fill_rule = "evenodd"),
       fill = "magenta", stroke = "black", stroke_width = 4)
}
}
\seealso{
\code{\link[=POLYGON]{POLYGON()}} and \code{\link[=ISOTOXAL_2NGON]{ISOTOXAL_2NGON()}}.
}
