% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_helpers.R
\name{x_ellipse_left}
\alias{x_ellipse_left}
\alias{x_ellipse_right}
\alias{y_ellipse_top}
\alias{y_ellipse_bottom}
\title{Find ellipse coordinate}
\usage{
x_ellipse_left(
  y,
  xc,
  yc,
  rx,
  ry = rx,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE)
)

x_ellipse_right(
  y,
  xc,
  yc,
  rx,
  ry = rx,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE)
)

y_ellipse_top(
  x,
  xc,
  yc,
  rx,
  ry = rx,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE)
)

y_ellipse_bottom(
  x,
  xc,
  yc,
  rx,
  ry = rx,
  a = 0,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE)
)
}
\arguments{
\item{y}{The y value at which to evaluate the ellipse.}

\item{xc, yc}{The center of the ellipse.}

\item{rx, ry}{The x and y radii of the (unrotated) ellipse.}

\item{a}{The counter-clockwise angle of rotation of the ellipse (will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}).}

\item{...}{Should be empty.}

\item{origin_at_bottom}{If \code{TRUE}, the origin is at the bottom of the
coordinate system (i.e. y increases upward).}

\item{x}{The x value at which to evaluate the ellipse.}
}
\value{
A numeric value.
}
\description{
\code{x_ellipse_left()} returns the x-coordinate of the leftmost point and
\code{x_ellipse_right()} returns the x-coordinate of the rightmost point on a
(possibly rotated) ellipse at a given y value.
\code{y_ellipse_top()} returns the y-coordinate of the topmost point and
\code{y_ellipse_bottom()} returns the y-coordinate of the bottommost point on a
(possibly rotated) ellipse at a given x value.
\code{x_ellipse_left()} and \code{x_ellipse_right()} return \code{NaN} if \code{y} is
outside the vertical extent of the ellipse.
\code{y_ellipse_top()} and \code{y_ellipse_bottom()} return \code{NaN} if \code{x} is
outside the horizontal extent of the ellipse.
}
\examples{
# Unrotated ellipse centered at (5, 5) with rx=3, ry=4
x_ellipse_left(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 0)
x_ellipse_right(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 0)
# Rotated 90 degrees: rx and ry swap roles
x_ellipse_left(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 90)
x_ellipse_right(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 90)
# Unrotated ellipse centered at (5, 5) with rx=3, ry=4
y_ellipse_top(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 0)
y_ellipse_bottom(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 0)
# Rotated 90 degrees: rx and ry swap roles
y_ellipse_top(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 90)
y_ellipse_bottom(5, xc = 5, yc = 5, rx = 3, ry = 4, a = 90)
}
\seealso{
\code{\link[=A]{A()}} and \code{\link[=d_ellipse]{d_ellipse()}}.
}
