% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{A}
\alias{A}
\alias{aa}
\alias{AZ}
\alias{az}
\title{The elliptical arc curve commands}
\usage{
A(
  rx,
  ry = rx,
  x_axis_rotation = 0,
  large_arc_flag = FALSE,
  sweep_flag = FALSE,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

aa(
  rx,
  ry = rx,
  x_axis_rotation = 0,
  large_arc_flag = FALSE,
  sweep_flag = FALSE,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

AZ(...)

az(...)
}
\arguments{
\item{rx, ry}{Radius of ellipse.}

\item{x_axis_rotation}{Angle (in degrees) from x-axis of ellipse.
Will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}.
If \code{isTRUE(origin_at_bottom)} will multiply by \code{-1}.}

\item{large_arc_flag}{If \code{TRUE} then one of two larger arc sweeps chosen else
one of the two smaller arc sweeps.}

\item{sweep_flag}{If \code{TRUE} then arc will be drawn in "positive-angle" direction.
else drawn in "negative-angle" direction.
If \code{isTRUE(origin_at_bottom)} will invert.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by \code{height - y}.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{A()} and \code{aa()} draw elliptical arc curve commands.
}
\examples{
M(1, 1) + A(rx = 1, x = 2, y = 2) + Z()
M(1, 1) + aa(rx = 1, x = 1, y = 1) + zz()
}
