% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{A}
\alias{A}
\alias{aa}
\alias{AZ}
\alias{az}
\title{The elliptical arc curve commands}
\usage{
A(
  rx,
  ry = rx,
  x_axis_rotation = -a,
  large_arc_flag = big,
  sweep_flag = cw,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf),
  a = 0,
  cw = FALSE,
  big = FALSE
)

aa(
  rx,
  ry = rx,
  x_axis_rotation = 0,
  large_arc_flag = FALSE,
  sweep_flag = FALSE,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

AZ(...)

az(...)
}
\arguments{
\item{rx, ry}{Radius of ellipse.}

\item{x_axis_rotation, a}{Angle (in degrees) to rotate ellipse clockwise.
Will be coerced by \code{\link[affiner:angle]{affiner::degrees()}}.
If \code{isTRUE(origin_at_bottom)} then will be multiplied by \code{-1} in the output.
\code{a} is an alternative that instead rotates the ellipse counter-clockwise.}

\item{large_arc_flag, big}{If \code{TRUE} then the arc will be one of two larger arc sweeps
else one of the two smaller arc sweeps. \code{big} is an alias.}

\item{sweep_flag, cw}{If \code{TRUE} then the arc will be one of the two possible arcs in the "clockwise" direction.
else one of the two in a "counter-clockwise" direction.
If \code{isTRUE(origin_at_bottom)} then will be inverted in the output.
\code{cw} is an alias.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}

\item{digits}{\code{x} and \code{y} (and any \code{x1}, \code{y1}, \code{x2}, \code{y2}) will be transformed by \code{round(, digits)}.
An \code{Inf} (default) means no rounding.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{A()} and \code{aa()} draw elliptical arc curve commands.
}
\examples{
M(1, 1) + A(rx = 1, x = 2, y = 2) + Z()
M(1, 1) + aa(rx = 1, x = 1, y = 1) + zz()
d_small_ccw <- M(40, 70) + A(rx = 20, ry = 30, x = 20, y = 20)
d_small_cw <- M(40, 70) + A(rx = 20, ry = 30, x = 20, y = 20, cw = TRUE)
d_big_ccw <- M(40, 70) + A(rx = 20, ry = 30, x = 20, y = 20, big = TRUE)
d_big_cw <- M(40, 70) + A(rx = 20, ry = 30, x = 20, y = 20, big = TRUE, cw = TRUE)
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(c(d_small_ccw, d_small_cw, d_big_ccw, d_big_cw),
    height = 100, width = 100,
    stroke = c("black", "grey", "blue", "cyan"),
    fill = "none", stroke_width = 4)
}
}
