% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{L}
\alias{L}
\alias{ll}
\alias{LZ}
\alias{lz}
\alias{H}
\alias{hh}
\alias{HZ}
\alias{hz}
\alias{V}
\alias{vv}
\alias{VZ}
\alias{vz}
\title{The "lineto" commands}
\usage{
L(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

ll(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

LZ(...)

lz(...)

H(x, ..., digits = getOption("dee.digits", Inf))

hh(x, ..., digits = getOption("dee.digits", Inf))

HZ(...)

hz(...)

V(
  y,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

vv(
  y,
  ...,
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

VZ(...)

vz(...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}

\item{digits}{\code{x} and \code{y} (and any \code{x1}, \code{y1}, \code{x2}, \code{y2}) will be transformed by \code{round(, digits)}.
An \code{Inf} (default) means no rounding.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{L()} and \code{ll()} draw straight lines,
\code{H()} and \code{hh()} draw horizontal lines, and
\code{V()} and \code{vv()} draw vertical lines.
}
\examples{
M(1, 1) + L(2, 2) + Z()
M(1, 1) + ll(1, 1) + zz()
d <- M(1,1) + L(5,8) + H(8) + V(3) + Z()
if (requireNamespace("omsvg", quietly = TRUE) &&
    requireNamespace("svgparser", quietly = TRUE)) {
  plot(d, height = 10, width = 10,
       fill = "red", stroke = "black", stroke_width = 4)
}
}
