% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{L}
\alias{L}
\alias{ll}
\alias{LZ}
\alias{lz}
\alias{H}
\alias{hh}
\alias{HZ}
\alias{hz}
\alias{V}
\alias{vv}
\alias{VZ}
\alias{vz}
\title{The "lineto" commands}
\usage{
L(x, y = NULL, sep = getOption("dee.sep", ","))

ll(x, y = NULL, sep = getOption("dee.sep", ","))

LZ(...)

lz(...)

H(x)

hh(x)

HZ(...)

hz(...)

V(y)

vv(y)

VZ(...)

vz(...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{...}{Passed to related function.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{L()} and \code{ll()} draw straight lines,
\code{H()} and \code{hh()} draw horizontal lines, and
\code{V()} and \code{vv()} draw vertical lines.
}
