% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{M}
\alias{M}
\alias{mm}
\alias{MZ}
\alias{mz}
\title{The "moveto" commands}
\usage{
M(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

mm(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL),
  digits = getOption("dee.digits", Inf)
)

MZ(...)

mz(...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}

\item{digits}{\code{x} and \code{y} (and any \code{x1}, \code{y1}, \code{x2}, \code{y2}) will be transformed by \code{round(, digits)}.
An \code{Inf} (default) means no rounding.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{M()} and \code{mm()} move the "pen" to a new point.
}
\examples{
M(1, 1) + L(2, 2) + Z()
M(1, 1) + ll(1, 1) + zz()
}
