% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{M}
\alias{M}
\alias{mm}
\alias{MZ}
\alias{mz}
\title{The "moveto" commands}
\usage{
M(x, y = NULL, sep = getOption("dee.sep", ","))

mm(x, y = NULL, sep = getOption("dee.sep", ","))

MZ(...)

mz(...)
}
\arguments{
\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{...}{Passed to related function.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{M()} and \code{mm()} move the "pen" to a new point.
}
