% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{Q}
\alias{Q}
\alias{qq}
\alias{QZ}
\alias{qz}
\alias{T}
\alias{tt}
\alias{TZ}
\alias{tz}
\title{The quadratic Bezier curve commands}
\usage{
Q(x1, y1 = NULL, x, y = NULL, sep = getOption("dee.sep", ","))

qq(x1, y1 = NULL, x, y = NULL, sep = getOption("dee.sep", ","))

QZ(...)

qz(...)

T(x, y = NULL, sep = getOption("dee.sep", ","))

tt(x, y = NULL, sep = getOption("dee.sep", ","))

TZ(...)

tz(...)
}
\arguments{
\item{x1}{If \code{y1} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y1}{Either \code{NULL} or a numeric vector.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{...}{Passed to related function.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{Q()} and \code{qq()} draw quadratic Bezier curves
\code{T()} and \code{tt()} draw quadratic Bezier curves assuming the control point is the reflection of the previous Bezier curve command.
}
