% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{Q}
\alias{Q}
\alias{qq}
\alias{QZ}
\alias{qz}
\alias{T}
\alias{tt}
\alias{TZ}
\alias{tz}
\title{The quadratic Bézier curve commands}
\usage{
Q(
  x1,
  y1 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

qq(
  x1,
  y1 = NULL,
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

QZ(...)

qz(...)

T(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

tt(
  x,
  y = NULL,
  ...,
  sep = getOption("dee.sep", ","),
  origin_at_bottom = getOption("dee.origin_at_bottom", FALSE),
  height = getOption("dee.height", NULL)
)

TZ(...)

tz(...)
}
\arguments{
\item{x1}{If \code{y1} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y1}{Either \code{NULL} or a numeric vector.}

\item{x}{If \code{y} is \code{NULL} will be coerced by \code{\link[affiner:as_coord2d]{affiner::as_coord2d()}}.
Else a numeric vector.}

\item{y}{Either \code{NULL} or a numeric vector.}

\item{...}{Passed to related function.}

\item{sep}{Either \code{","} or \code{" "}.}

\item{origin_at_bottom, height}{If \code{origin_at_bottom} is \code{TRUE} then
\code{y} (and any \code{y1} and \code{y2}) coordinates is transformed by
\code{height - y} for absolute coordinates and \code{-y} for relative coordinates.}
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{Q()} and \code{qq()} draw quadratic Bézier curves
\code{T()} and \code{tt()} draw quadratic Bézier curves assuming the control point is the reflection of the previous Bézier curve command.
}
\examples{
M(1, 1) + Q(2, 2, 3, 3) + T(4, 4) + Z()
M(1, 1) + qq(1, 1, 2, 2) + tt(1, 1) + zz()
}
