% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{Z}
\alias{Z}
\alias{zz}
\title{The "closepath" commands}
\usage{
Z()

zz()
}
\value{
A \code{\link[=dee]{dee()}} object.
}
\description{
\code{Z()} and \code{zz()} connects the initial point of the current subpath with the last point (with a straight line if these are different).  There is no difference between \code{Z()} and \code{zz()}.
}
\examples{
M(1, 1) + L(2, 2) + Z()
M(1, 1) + ll(2, 2) + zz()
}
